pyproject.toml
[tool.poetry]
name = "new-admin-panel-sprint-3"
version = "0.1.0"
description = "..."
authors = ["Aleksey Gredyaev <aleksey.gredyaev@gmail.com>"]
readme = "README.md"

# Dependencies
[tool.poetry.dependencies]
python = ">=3.9, <4.0"

[tool.poetry.group.backend.dependencies]
psycopg2-binary = "^2.9"
django = "^4.2.5"
python-dotenv = "^1.0.1"
django-split-settings = "^1.2.0"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.1"
django-extensions = "^3.2.3"

[tool.poetry.group.etl.dependencies]
elasticsearch = "^8.12.1"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
pydantic = "^2.7.0"
redis = "^5.0.3"

[tool.poetry.group.dev.dependencies]
flake8 = "5.0.4"
flake8-blind-except = "0.2.1"
flake8-broken-line = "0.6.0"
flake8-bugbear = "22.9.23"
flake8-builtins = "2.0.0"
flake8-class-attributes-order = "0.1.3"
flake8-cognitive-complexity = "0.1.0"
flake8-commas = "2.1.0"
flake8-comprehensions = "3.10.0"
flake8-debugger = "4.1.2"
flake8-eradicate = "1.4.0"
flake8-functions = "0.0.7"
flake8-isort = "5.0.0"
flake8-mutable = "1.2.0"
flake8-print = "5.0.0"
flake8-pytest = "1.4"
flake8-pytest-style = "1.6.0"
flake8-quotes = "3.3.1"
flake8-string-format = "0.3.0"
flake8-variables-names = "0.0.5"
isort = "5.10.1"
pytest = "7.1.1"
pytest-cov = "4.0.0"
Faker = "20.0.0"
pre-commit = "3.6.0"
pytest-flake8 = "1.0.0"
autopep8 = "^1.5.7"
yapf = "^0.31.0"
click = "^8.1.7"
black = "^24.2.0"
pylint-django = "^2.5.5"
requests = "^2.31.0"
jsonschema = "^4.21.1"
ruff = "^0.4.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.flake8]
max-line-length = 88
exclude = [
  ".git",
  "__pycache__",
  "build",
  "dist",
  "*.pyc",
  "*.egg-info",
  ".venv",
  ".eggs",
  ".tox",
]
ignore = ["W503"]
max-complexity = 10

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -v"
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
